+++ src/pages/seokyoung/Main/comment/Comment.js
@@ -1,11 +1,12 @@
 import React from "react";
 
-function Comment({ comment }) {
+function Comment({ comment, deleteFunction }) {
+  // console.log(deleteFunction);
   return (
     <li className="commentPost">
       <span>{comment.name}</span>
       <div>{comment.text}</div>
-      <i className="fa-solid fa-xmark deleteBtn" />
+      <i onClick={deleteFunction} className="fa-solid fa-xmark deleteBtn" />
     </li>
   );
 }
