+++ src/pages/seokyoung/Main/comments/Comments.js
@@ -1,13 +1,11 @@
 import React from "react";
 import Comment from "../comment/Comment";
 
-function Comments({ commentsList }) {
-  console.log(commentsList);
-
+function Comments({ commentsList, deleteComment }) {
   return (
     <ul className="commentPostList">
-      {commentsList.map((item, i) => (
-        <Comment key={i} comment={item} />
+      {commentsList.map(item => (
+        <Comment key={item.id} comment={item} deleteFunction={deleteComment} />
       ))}
     </ul>
   );
