+++ src/pages/seokyoung/Main/mainFeed/feedPost/FeedPost.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+const FeedPost = () => {
+  return (
+    <>
+      <div className="feedTop">
+        <img src="images/seokyoung/feed.jpeg" alt="" />
+        <div className="feedTopName">heaundea_DDol</div>
+      </div>
+      <img className="feedImg" src="images/seokyoung/ddol.JPG" alt="" />
+    </>
+  );
+};
+
+export default FeedPost;
