+++ src/pages/seokyoung/Main/rightFeed/friends/friend/Friend.js
@@ -1,0 +1,40 @@
+import React from "react";
+
+function Friend() {
+  return (
+    <div className="freind1">
+      <img src="images/seokyoung/freind1.png" alt="친구1" />
+      <div className="profileName">
+        wecode_bootCamp
+        <div className="profileKor">위코드 32기</div>
+      </div>
+      <div className="profileChange">팔로우</div>
+    </div>
+  );
+}
+
+// 프렌드 2
+{
+  /* <div className="freind1">
+<img src="images/seokyoung/freind2.png" alt="친구2" />
+<div className="profileName">
+  finda_offical
+  <div className="profileKor">핀다</div>
+</div>
+<div className="profileChange">팔로우</div>
+</div> */
+}
+
+// 프렌드 3
+{
+  /* <div className="freind1">
+<img src="images/seokyoung/friend3.png" alt="친구3" />
+<div className="profileName">
+  kakao_developer
+  <div className="profileKor">카카오</div>
+</div>
+<div className="profileChange">팔로우</div>
+</div> */
+}
+
+export default Friend;
